requirements.txt
google-cloud-storage
google-cloud-firestore
moviepy
functions-framework>=3.0.0